requirements.txt
google-genai>=0.8.0
faiss-cpu>=1.7.3
protobuf>=4.21.0
google-cloud-storage>=2.11.0
